Sprint 754: relocate client-access helpers (deferred-items condition met)
The 2026-04-20 deferred-items entry said: "revisit only if a fourth
helper joins them." Sprint 735 added require_client_owner as the
fourth, so this sprint makes the move.

shared/client_access.py (new, 100 lines): is_authorized_for_client,
get_accessible_client, require_client, require_client_owner. Module
docstring documents the org-scoped vs direct-only access policy.

shared/helpers.py: client-access definitions removed; module is now
pure parsing/coercion utilities (try_parse_risk*, parse_json_*).

All 6 callers migrated to import from shared.client_access directly:
- routes/clients.py
- routes/diagnostics.py
- routes/prior_period.py
- routes/settings.py
- routes/trends.py
- tests/test_sprint_735_require_client_owner.py

No shim on shared.helpers — matches Sprint 724's discipline of
avoiding re-export shims.

test_no_helpers_reexports ALLOWED_HELPER_NAMES shrunk to the 4 parsing
helpers; the guardrail continues to prevent the shim from being
re-introduced. test_refactor_2026_04_20 imports split between the two
owning modules.

70/70 affected tests pass (guardrail + 51 consumer-route tests across
clients/diagnostics/prior_period/settings/trends APIs).

Sprint 754's other items (shared analysis interfaces, broader helper
dedup, module-boundary tests) deferred to follow-ups — design work
beyond one-sprint scope, and not pre-requisites for Phases 6-8.
